move create_mesh to fem gs utilities
